+++ config/routes.rb
@@ -242,6 +242,7 @@ Rails.application.routes.draw do
       scope only: :index do
         resources :plans, only: %i[index show]
         resources :plan_categories
+        resources :prices
         resources :users
         resources :trainings
         resources :user_trainings