+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - OpenAPI spaces endpoints (index/show)
 - OpenAPI plans endpoints (index/show)
 - OpenAPI plans categories index endpoint
+- OpenAPI prices index endpoint
 - Improved automated test on statistics generation
 - Refactored statistics generation
 - Refactored test helpers